FirstExample

+++ src/com/epam/tcoi/FirstExample.java
@@ -2,7 +2,7 @@ package com.epam.tcoi;
 
 public class FirstExample {
     public static void main(String[] args) {
-        int arrayLength = 0;
+        double arrayLength = 0;
         for (String arg : args) arrayLength += arg.length();
         System.out.println(args.length);
         System.out.println(arrayLength/args.length);
